* Make slice for Video conversations
